Move function (it's used in patch, too)
